[BACKEND] Don't crash on an assert when using M=64 & 2CTA mode
The current code does not work yet as it does not play ball with the
recent `CGALayout` refactor. If we fix this I'd hope we should be able
to make it work, modulo the same limitations that `M=128` has. We will
decide at a later point whether we want to invest on this or not.